refactor(7/inroduce): 섹션 구조 분리 및 모듈화 정리
- _components 평면 구조를 _animation/_layout/_sections로 재구성
- Staff, Final, Projects, Intro 애니메이션 로직을 컴포넌트/훅 단위로 분리
- 배럴(index.ts)과 structure.md를 추가해 import 경로와 구조 문서를 간결하게 정리
