+++ src/app/7/introduce/_animation/slot.constants.ts
@@ -1,0 +1,3 @@
+export const REEL_DIGITS = [...Array.from({ length: 30 }, (_, i) => i % 10), 7];
+export const REEL_POSITIONS = ['left', 'center', 'right'] as const;
+
