+++ src/app/7/introduce/_layout/CurriculumGlass.module.css
