+++ src/app/7/introduce/_sections/curriculum/PartPreviewLottie.tsx
