+++ src/app/7/introduce/_sections/curriculum/curriculum.types.ts
