+++ src/app/7/introduce/_sections/introduce/index.ts
@@ -1,0 +1,2 @@
+export { default as IntroduceSection } from './IntroduceSection';
+
