+++ src/app/7/introduce/_sections/projects/index.ts
@@ -1,0 +1,6 @@
+export { default as ProjectsSection } from './ProjectsSection';
+export { default as ProjectCard } from './ProjectCard';
+export { default as ProjectMarqueeRow } from './ProjectMarqueeRow';
+export { default as useProjectRows } from './useProjectRows';
+export type { ProjectItem, ProjectRow } from './project.types';
+
