+++ src/app/7/introduce/_sections/projects/project.types.ts
@@ -1,0 +1,11 @@
+export type ProjectItem = {
+  batch: string;
+  name: string;
+  description: string;
+};
+
+export type ProjectRow = {
+  batch: string;
+  items: ProjectItem[];
+};
+
