+++ src/app/7/introduce/_sections/staff/index.ts
@@ -1,0 +1,6 @@
+export { default as StaffSection } from './StaffSection';
+export { default as StaffCard } from './StaffCard';
+export { default as StaffSkeletonGrid } from './StaffSkeletonGrid';
+export { default as StaffSectionTemplate } from './StaffSectionTemplate';
+export type { StaffItem, StaffPart, StaffTab } from './staff.types';
+
