+++ src/app/7/introduce/_sections/staff/staff.types.ts
@@ -1,0 +1,15 @@
+export type StaffPart = 'management' | 'android' | 'web' | 'server' | 'pm' | 'design';
+
+export type StaffItem = {
+  part: StaffPart;
+  name: string;
+  image: string;
+  comment: string;
+  role: string;
+};
+
+export type StaffTab = {
+  key: StaffPart;
+  label: string;
+};
+
